+++ artifacts/PowerShell/Bootstrap.ps1
@@ -17,7 +17,8 @@ param (
   [string]$autoDeployClusterResource,
   [string]$autoUpgradeClusterResource,
   [string]$debugEnabled,
-  [string]$vmAutologon
+  [string]$vmAutologon,
+  [string]$clusterNodeCount = '3'
 )
 
 Write-Output "Input parameters:"
@@ -38,6 +39,7 @@ $PSBoundParameters
 [System.Environment]::SetEnvironmentVariable('autoUpgradeClusterResource', $autoUpgradeClusterResource, [System.EnvironmentVariableTarget]::Machine)
 [System.Environment]::SetEnvironmentVariable('registerCluster', $registerCluster, [System.EnvironmentVariableTarget]::Machine)
 [System.Environment]::SetEnvironmentVariable('natDNS', $natDNS, [System.EnvironmentVariableTarget]::Machine)
+[System.Environment]::SetEnvironmentVariable('clusterNodeCount', $clusterNodeCount, [System.EnvironmentVariableTarget]::Machine)
 [System.Environment]::SetEnvironmentVariable('LocalBoxDir', "C:\LocalBox", [System.EnvironmentVariableTarget]::Machine)
 [System.Environment]::SetEnvironmentVariable('LocalBoxLogsDir', "C:\LocalBox\Logs", [System.EnvironmentVariableTarget]::Machine)
 [System.Environment]::SetEnvironmentVariable('LocalBoxTestsDir', "C:\LocalBox\Tests", [System.EnvironmentVariableTarget]::Machine)